refactor(automations): slackWebApi becomes a package factory client (ADR-063 P1)
createSlackWebApiClient({ egress }) moves to
automations-server/clients/slack/web-api.client with chat.postMessage,
conversations.list paging, and the scope-degradation fallback intact; the
app constructs one instance in delivery/appSlackWebApi over its SSRF
egress, so all six consumers keep their call signatures. The unit suite
moves into the package driving the factory with a fake egress.
