+++ langwatch/src/server/api/routers/automations.ts
@@ -18,7 +18,7 @@ import {
 import { getApp } from "~/server/app-layer/app";
 import { HandledError } from "@langwatch/handled-error";
 import { translateFilterToClickHouse } from "~/server/app-layer/traces/filter-to-clickhouse";
-import { listSlackChannels } from "~/server/app-layer/automations/delivery/slackWebApi";
+import { listSlackChannels } from "~/server/app-layer/automations/delivery/appSlackWebApi";
 import {
   InvalidEmailRecipientError,
   MissingAnnotatorError,
