+++ langwatch/src/server/app-layer/automations/delivery/triggerNotifier.ts
@@ -6,7 +6,7 @@ import type { TriggerNotifier } from "~/server/app-layer/automations/trigger-tem
 import { sendEmail } from "~/server/mailer/emailSender";
 import { assertWebhookDelivered, sendWebhook } from "./sendWebhook";
 import { isSlackWebhookUrl } from "@langwatch/automations-server/clients/slack/webhook-guard";
-import { postSlackChatMessage } from "./slackWebApi";
+import { postSlackChatMessage } from "./appSlackWebApi";
 
 /**
  * Production delivery for trigger test fires: the email path reuses the shared
