+++ langwatch/src/server/app-layer/reports/report-dispatch.ts
@@ -10,7 +10,7 @@ import {
 } from "@langwatch/automations/providers/slack";
 import type { sendRenderedTriggerEmail } from "~/server/mailer/triggerEmail";
 import type { sendRenderedSlackMessage } from "~/server/app-layer/automations/delivery/sendSlackWebhook";
-import type { postSlackChatMessage } from "~/server/app-layer/automations/delivery/slackWebApi";
+import type { postSlackChatMessage } from "~/server/app-layer/automations/delivery/appSlackWebApi";
 import { REPORT_TRIGGER_DEFAULTS } from "@langwatch/automations/templating/defaults";
 import { renderTriggerEmail } from "@langwatch/automations/templating/renderEmail";
 import {
